requirements.txt
disnake==2.5.1
python_dateutil==2.8.2
PyYAML==6.0
